requirements.txt
sxtwl
python-dateutil>=2.8.2
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
langchain>=0.3.15
langchain-ollama
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langgraph
rich
pypdf
notebook
python-dotenv==1.0.0
requests==2.31.0